fix: UI: Split expression-util out of mapping-management.service
Fixes: #2881
fix: UI: Move document-management.service#setLibraryToService() to file-management.service
Fixes: #2880
Also partially addresses #2726, isolated digest file handling into mapping-digest-util
