+++ ui/packages/atlasmap-core/src/common/constants.ts
@@ -13,5 +13,6 @@
     See the License for the specific language governing permissions and
     limitations under the License.
 */
-export * from './config.types';
-export * from './constants';
+
+export const HTTP_STATUS_OK = 200;
+export const HTTP_STATUS_NO_CONTENT = 204;
