+++ ui/packages/atlasmap-core/src/utils/index.ts
@@ -13,8 +13,11 @@
     See the License for the specific language governing permissions and
     limitations under the License.
 */
+export * from './common-util';
+export * from './expression-util';
 export * from './filter-fields';
 export * from './guid';
 export * from './lookup-table-util';
+export * from './mapping-digest-util';
 export * from './mapping-serializer';
 export * from './mapping-util';
