+++ ui/packages/atlasmap-core/src/utils/mapping-digest-util.spec.ts
@@ -1,0 +1,44 @@
+/*
+    Copyright (C) 2017 Red Hat, Inc.
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+            http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+*/
+
+import { DocumentDefinition } from '../models';
+import { MappingDigestUtil } from './mapping-digest-util';
+
+describe('MappingDigestUtil', () => {
+  test('generateExportMappings()', () => {
+    expect(MappingDigestUtil.generateExportMappings('')).toBe('');
+    const answer = MappingDigestUtil.generateExportMappings(
+      '{"dummyMapping": {}}'
+    );
+    expect(answer).toContain('exportMappings');
+    expect(answer).toContain('"value": "{\\"dummyMapping\\": {}}"');
+  });
+
+  test('generateExportBlockData()', () => {
+    expect(MappingDigestUtil.generateExportBlockData('')).toBe('');
+    const answer = MappingDigestUtil.generateExportBlockData('{"dummy": {}}');
+    expect(answer).toContain('"value": "{\\"dummy\\": {}}"');
+  });
+
+  test('generateExportMetaStr()', () => {
+    const doc = new DocumentDefinition();
+    doc.name = 'dummyDoc';
+    doc.id = 'dommyDocId';
+    const answer = MappingDigestUtil.generateExportMetaStr(doc);
+    expect(answer).toContain(doc.name);
+    expect(answer).toContain(doc.id);
+  });
+});
