+++ ui/packages/atlasmap-core/src/utils/mapping-digest-util.ts
@@ -1,0 +1,75 @@
+/*
+    Copyright (C) 2017 Red Hat, Inc.
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+            http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+*/
+import { CommonUtil } from './common-util';
+import { DocumentDefinition } from '../models/document-definition.model';
+
+export class MappingDigestUtil {
+  /**
+   * Capture the specified user mappings into a general catalog JSON buffer (exportMappings).
+   * @param buffer
+   */
+  static generateExportMappings(buffer: string): string {
+    if (!buffer || buffer.length === 0) {
+      return '';
+    }
+    return (
+      `   "exportMappings":
+    {
+       "value": ` +
+      CommonUtil.sanitizeJSON(buffer) +
+      `
+    },\n`
+    );
+  }
+
+  /**
+   * Capture the specified user JSON or XML document buffer into a general catalog JSON buffer.
+   *
+   * @param buffer
+   */
+  static generateExportBlockData(buffer: string): string {
+    if (buffer === null || buffer.length === 0) {
+      return '';
+    }
+    return (
+      `
+          {
+             "value": ` +
+      CommonUtil.sanitizeJSON(buffer) +
+      `
+          }`
+    );
+  }
+
+  /**
+   * Capture the specified user document definition meta data into a general
+   * catalog JSON buffer.
+   *
+   * @param docDef
+   */
+  static generateExportMetaStr(docDef: DocumentDefinition): string {
+    const inspectionParameters = JSON.stringify(docDef.inspectionParameters);
+    return `
+       {
+          "name": "${docDef.name}",
+          "dataSourceType": "${docDef.type}",
+          "id": "${docDef.id}",
+          "inspectionType": "${docDef.inspectionType}",
+          "inspectionParameters": ${inspectionParameters},
+          "isSource": "${docDef.isSource}"
+       }`;
+  }
+}
